requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
bcrypt
jwt
httpx
supabase
google-genai
